[FLINK-16504] [dynamodb connector] move WriteRequest and WriteResponse
