+++ flink-connectors/flink-connector-dynamodb/src/main/java/org/apache/flink/streaming/connectors/dynamodb/WriteResponse.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.connectors.dynamodb.batch;
+package org.apache.flink.streaming.connectors.dynamodb;
 
 /** DynamoDB batch writer response. */
 public class WriteResponse {
