+++ flink-connectors/flink-connector-dynamodb/src/main/java/org/apache/flink/streaming/connectors/dynamodb/batch/BatchCollector.java
@@ -18,6 +18,7 @@
 
 package org.apache.flink.streaming.connectors.dynamodb.batch;
 
+import org.apache.flink.streaming.connectors.dynamodb.WriteRequest;
 import org.apache.flink.streaming.connectors.dynamodb.batch.key.PrimaryKey;
 import org.apache.flink.streaming.connectors.dynamodb.config.DynamoDbTablesConfig;
