+++ flink-connectors/flink-connector-dynamodb/src/test/java/org/apache/flink/streaming/connectors/dynamodb/DynamoDBBatchWriterTest.java
@@ -20,7 +20,6 @@ package org.apache.flink.streaming.connectors.dynamodb;
 
 import org.apache.flink.streaming.connectors.dynamodb.batch.BatchWriterAttemptResult;
 import org.apache.flink.streaming.connectors.dynamodb.batch.DynamoDbBatchWriter;
-import org.apache.flink.streaming.connectors.dynamodb.batch.WriteResponse;
 import org.apache.flink.streaming.connectors.dynamodb.retry.BatchWriterRetryPolicy;
 import org.apache.flink.streaming.connectors.dynamodb.retry.DefaultBatchWriterRetryPolicy;
